requirements.txt
requests
supabase
python-dotenv
keyring
numpy
pandas
yfinance
